+++ test/numeric/num_test.dart
@@ -3,7 +3,7 @@ import 'package:formdator/formdator.dart';
 
 void main() {
   const error = 'The value must be numeric';
-  final numeric = Num(nan: error);
+  final numeric = Num(non: error);
 
   group('Num', () {
     group('():', () {
@@ -47,7 +47,7 @@ void main() {
     });
     group('min:', () {
       const small = 'too small';
-      final numMin = Num.min(10, nan: error, small: small);
+      final numMin = Num.min(10, non: error, small: small);
       test('null input', () {
         expect(numMin(null), null);
       });
@@ -66,7 +66,7 @@ void main() {
     });
     group('max:', () {
       const large = 'too large';
-      final numMax = Num.max(10, nan: error, large: large);
+      final numMax = Num.max(10, non: error, large: large);
       test('null', () {
         expect(numMax(null), null);
       });
@@ -87,9 +87,9 @@ void main() {
     group('range:', () {
       const small = 'too small';
       const large = 'too large';
-      final range = Num.range(1, 10, nan: error, small: small, large: large);
+      final range = Num.range(1, 10, non: error, small: small, large: large);
       final negRange =
-          Num.range(-10, -1, nan: error, small: small, large: large);
+          Num.range(-10, -1, non: error, small: small, large: large);
       test('null', () {
         expect(range(null), null);
         expect(negRange(null), null);